Add tidying example
A quick example of how to tidy the veg height data.

It's more complicated than strictly needs to be because I was writing it thinking about the larger data set it belongs to. The other Excel workbooks from that data source don't all perfectly follow the same format, so this code compensates for that.